Dialogs and Toasts

+++ app/favorites/favorites.component.ts
@@ -5,6 +5,8 @@ import { ListViewEventData, RadListView } from 'nativescript-ui-listview';
 import { RadListViewComponent } from 'nativescript-ui-listview/angular';
 import { ObservableArray } from 'tns-core-modules/data/observable-array';
 import { View } from 'tns-core-modules/ui/core/view';
+import { confirm } from 'ui/dialogs';
+import { Toasty } from 'nativescript-toasty';
 
 @Component({
     selector: 'app-favorites',
@@ -30,9 +32,29 @@ export class FavoritesComponent implements OnInit{
     }
 
     deleteFavorite(id: number) {
-        this.favoriteService.deleteFavorite(id)
-            .subscribe(favorites => this.favorites = new ObservableArray(favorites),
+        console.log('delete', id);
+
+        confirm({
+            title: 'Confirm Delete',
+            message: 'Do you want to delete Dish '+id,
+            okButtonText: 'Yes',
+            cancelButtonText: 'No',
+            neutralButtonText: 'Cancel'
+        }).then((result: boolean) => {
+            if(result) {
+                this.favorites = null;
+
+                this.favoriteService.deleteFavorite(id)
+                .subscribe(favorites => {
+                    const toast = new Toasty('Deleted Dish ' + id, 'short', 'bottom');
+                    toast.show();
+                    this.favorites = new ObservableArray(favorites);
+                },
                 errmess => this.errMess = <any>errmess);
+            } else {
+                console.log('Deleted cancelled');
+            }
+        });
     }
 
     public onCellSwiping(args: ListViewEventData) {
